+++ CHANGELOG.md
@@ -19,6 +19,7 @@ with exception of a "0." prefix until version 1 is released.
 
  - Instruction `def` - LINE-MATCHER argument is mandatory
  - Instruction `def` - FILE-MATCHER argument is mandatory
+ - Instruction `def` - STRING-TRANSFORMER argument is mandatory
 
 ## [0.11.1.0]